+++ src/platforms/wordpress/Framework/Page.php
@@ -12,6 +12,23 @@ namespace Gantry\Framework;
 
 class Page extends Base\Page
 {
+    public $home;
+    public $outline;
+    public $language;
+    public $direction;
+
+    public function __construct($container)
+    {
+        parent::__construct($container);
+
+        $site = Gantry::instance()['site'];
+
+        $this->home = is_front_page();
+        $this->outline = Gantry::instance()['configuration'];
+        $this->language = (string) $site->language;
+        $this->direction = function_exists('is_rtl') && is_rtl() ? 'rtl' : 'ltr';
+    }
+
     public function url(array $args = [])
     {
         return home_url(add_query_arg($args, $GLOBALS['wp']->request));
@@ -19,17 +36,10 @@ class Page extends Base\Page
 
     public function htmlAttributes()
     {
-        $site = Gantry::instance()['site'];
-
-        $dir = [];
-        if(function_exists('is_rtl') && is_rtl()) {
-            $dir['dir'] = 'rtl';
-        }
-
         $attributes = [
-                'lang' => (string) $site->language,
+                'lang' => $this->language,
+                'dir' => $this->direction
               ]
-              + (array) $dir
               + (array) $this->config->get('page.html', []);
 
         return $this->getAttributes($attributes);
@@ -43,9 +53,9 @@ class Page extends Base\Page
         $body_classes = apply_filters('gantry5_body_classes', [
                 'site',
                 'outline-' . Gantry::instance()['configuration'],
+                'dir-' . $this->direction
             ]);
 
-        (is_rtl()) ? $body_classes[] = 'dir-rtl' : $body_classes[] = 'dir-ltr';
         $wp_body_class = get_body_class($body_classes);
 
         if(is_array($wp_body_class) && !empty($wp_body_class)) {