+++ CHANGELOG.md
@@ -22,6 +22,9 @@
     3. [](#bugfix)
         - Fixed `Joomla Articles` particle limits category selection to 20 categories only (thanks @nikola3244 - #1260)
         - Fixed `The menu selected is empty!` in Menu editor when PHP `default_charset` is not `UTF-8` (#1257)
+3. [WordPress](#wordpress)
+    2. [](#improved)
+        - Added missing `home`, `outline`, `language` and `direction` properties to `Framework\Page` class
         
 # 5.2.15
 ## 25/04/2016